backend/requirements.txt
Django==4.2.1
django-colorfield==0.8.0
django-cors-headers
django-extra-fields==3.0.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==6.0.0
gunicorn==20.0.4
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
